backend code quality and tests

+++ backend/src/main/java/org/solen/business/habitcases/DeleteHabitUseCaseImpl.java
@@ -1,6 +1,7 @@
 package org.solen.business.habitcases;
 
 import lombok.AllArgsConstructor;
+import org.solen.business.exceptions.HabitNotFoundByIdException;
 import org.solen.business.repos.IHabitRepository;
 import org.springframework.stereotype.Service;
 
@@ -13,6 +14,9 @@ public class DeleteHabitUseCaseImpl implements IDeleteHabitUseCase {
 
     @Override
     public void deleteHabit(Long id) {
+        if (repository.findById(id) == null) {
+            throw new HabitNotFoundByIdException(id);
+        }
         repository.deleteById(id);
     }
 }
